Major update: Triton removal and comprehensive GPU test suite
Changes:
- Permanently disabled Triton in all configurations (enable_triton=False)
- Removed Triton dependencies while preserving kernel files for reference
- Fixed all module imports to use correct paths (independent_core.*)
- Created comprehensive GPU test suite with hard failure mode
- Added test files for device management, memory management, and integration
- Fixed p-adic compression imports and relative imports
- Updated documentation to reflect Triton removal
- All components now work on CPU-only (Mac) and GPU (NVIDIA) systems

Test Suite Features:
- 110+ GPU tests covering all critical functionality
- Hard failure mode - no silent errors or fallbacks
- Memory leak detection and OOM handling tests
- Complete coverage of compression systems on GPU
- Master test runner with detailed reporting

All changes ensure backward compatibility and maintain functionality.

+++ independent_core/production_scaling/predictive_analytics.py
@@ -434,7 +434,7 @@ class PredictiveScalingAnalytics:
                 'components': {}
             }
             
-            time_points = range(0, time_range_hours * 60, 30)  # Every 30 minutes
+            time_points = range(0, int(time_range_hours * 60), 30)  # Every 30 minutes
             
             for component_name in self.monitor.get_all_system_status():
                 component_predictions = []
